backend/requirements.txt
# CRM Agent Backend - Python 3.11
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# LLM & Agents
openai>=1.57.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langgraph>=0.2.0

# Vector & Embeddings
chromadb>=0.5.0

# MCP
mcp>=1.0.0

# Data & HTTP
httpx>=0.28.0
datasets>=3.0.0

# Utilities
python-multipart>=0.0.17
